🧹 refactor packages for filesystem handling
