+++ providers/os/fsutil/find_files_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) Mondoo, Inc.
 // SPDX-License-Identifier: BUSL-1.1
 
-package fs
+package fsutil
 
 import (
 	"fmt"
