+++ providers/os/fsutil/find_files_windows.go
@@ -4,7 +4,7 @@
 //go:build windows
 // +build windows
 
-package fs
+package fsutil
 
 import (
 	"errors"
