+++ providers/os/mountedfs/file.go
@@ -1,7 +1,7 @@
 // Copyright (c) Mondoo, Inc.
 // SPDX-License-Identifier: BUSL-1.1
 
-package fs
+package mountedfs
 
 import (
 	"os"
