+++ providers/os/mountedfs/fs.go
@@ -1,7 +1,7 @@
 // Copyright (c) Mondoo, Inc.
 // SPDX-License-Identifier: BUSL-1.1
 
-package fs
+package mountedfs
 
 import (
 	"errors"
@@ -12,6 +12,7 @@ import (
 
 	"github.com/spf13/afero"
 	"go.mondoo.com/cnquery/v12/providers/os/connection/shared"
+	"go.mondoo.com/cnquery/v12/providers/os/fsutil"
 )
 
 var _ shared.FileSearch = (*MountedFs)(nil)
@@ -118,5 +119,5 @@ func (t *MountedFs) Chown(name string, uid, gid int) error {
 
 func (t *MountedFs) Find(from string, r *regexp.Regexp, typ string, perm *uint32, depth *int) ([]string, error) {
 	iofs := afero.NewIOFS(t)
-	return FindFiles(iofs, from, r, typ, perm, depth)
+	return fsutil.FindFiles(iofs, from, r, typ, perm, depth)
 }
